Fiberglass. Add KMP support for android and ios targets
