+++ composites-fiberglass/build.gradle.kts
@@ -16,14 +16,30 @@
 
 plugins {
     alias(libs.plugins.composites.library.compose)
+    alias(libs.plugins.composites.kotlin.multiplatform)
+    alias(libs.plugins.composites.compose.multiplatform)
     alias(libs.plugins.composites.spotless)
-    alias(libs.plugins.composites.dokka)
+    alias(libs.plugins.composites.dokka.multiplatform)
+    alias(libs.plugins.composites.maven.publish.multiplatform)
 }
 
 android {
     namespace = "ru.ldralighieri.composites.fiberglass"
 }
 
-dependencies {
-    implementation(libs.androidx.paging.compose)
+kotlin {
+    jvmToolchain {
+        languageVersion.set(JavaLanguageVersion.of(JavaVersion.VERSION_21.majorVersion))
+    }
+
+    sourceSets {
+        commonMain.dependencies {
+            api(compose.runtime)
+            api(compose.foundation)
+        }
+
+        androidMain.dependencies {
+            api(libs.androidx.paging.compose.android)
+        }
+    }
 }
