+++ composites-fiberglass/src/commonMain/kotlin/ru/ldralighieri/composites/fiberglass/column/FiberglassColumn.kt
