+++ composites-fiberglass/src/commonMain/kotlin/ru/ldralighieri/composites/fiberglass/column/FiberglassFlowColumn.kt
@@ -18,6 +18,7 @@ package ru.ldralighieri.composites.fiberglass.column
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.ExperimentalLayoutApi
 import androidx.compose.foundation.layout.FlowColumn
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.padding
@@ -41,6 +42,7 @@ import ru.ldralighieri.composites.fiberglass.model.FiberglassItem
  * @param maxItemsInEachColumn The maximum number of items per column.
  * @param itemKey A factory of the content types for the item.
  */
+@OptIn(ExperimentalLayoutApi::class)
 @Composable
 fun FiberglassFlowColumn(
     items: List<FiberglassItem>,
