+++ composites-fiberglass/src/commonMain/kotlin/ru/ldralighieri/composites/fiberglass/column/FiberglassLazyColumn.kt
@@ -29,9 +29,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import androidx.paging.compose.LazyPagingItems
-import androidx.paging.compose.itemContentType
-import androidx.paging.compose.itemKey
 import ru.ldralighieri.composites.fiberglass.model.FiberglassItem
 import ru.ldralighieri.composites.fiberglass.model.FiberglassLazyItemSlots
 import ru.ldralighieri.composites.fiberglass.model.FiberglassStickyHeaderItem
@@ -88,61 +85,6 @@ fun FiberglassLazyColumn(
             contentType = itemContentType,
         ) { position, item ->
             itemSlots[item::class]?.let { it(position, item) }
-        }
-    }
-}
-
-/**
- * Fiberglass lazy column Composite for [LazyPagingItems]
- *
- * @param items LazyPagingItems items.
- * @param itemSlots FiberglassLazyColumn [slots map][FiberglassLazyItemSlots]..
- * @param modifier The modifier to apply to this layout.
- * @param state The state object to be used to control or observe the list's state.
- * @param contentPadding A padding around the whole content.
- * @param reverseLayout Reverse the direction of scrolling and layout.
- * @param verticalArrangement The vertical arrangement of the layout's children.
- * @param horizontalAlignment The horizontal alignment applied to the items.
- * @param flingBehavior Logic describing fling behavior.
- * @param userScrollEnabled Whether the scrolling via the user gestures or accessibility actions
- * is allowed.
- * @param itemKey A factory of stable and unique keys representing the item.
- * @param itemContentType A factory of the content types for the item.
- */
-@Composable
-fun FiberglassLazyColumn(
-    items: LazyPagingItems<out FiberglassItem>,
-    itemSlots: FiberglassLazyItemSlots,
-    modifier: Modifier = Modifier,
-    state: LazyListState = rememberLazyListState(),
-    contentPadding: PaddingValues = PaddingValues(0.dp),
-    reverseLayout: Boolean = false,
-    verticalArrangement: Arrangement.Vertical =
-        if (!reverseLayout) Arrangement.Top else Arrangement.Bottom,
-    horizontalAlignment: Alignment.Horizontal = Alignment.Start,
-    flingBehavior: FlingBehavior = ScrollableDefaults.flingBehavior(),
-    userScrollEnabled: Boolean = true,
-    itemKey: ((item: FiberglassItem) -> Any)? = { it.id },
-    itemContentType: ((item: FiberglassItem) -> Any?)? = { it::class.simpleName },
-) {
-    LazyColumn(
-        modifier = modifier,
-        state = state,
-        contentPadding = contentPadding,
-        reverseLayout = reverseLayout,
-        verticalArrangement = verticalArrangement,
-        horizontalAlignment = horizontalAlignment,
-        flingBehavior = flingBehavior,
-        userScrollEnabled = userScrollEnabled,
-    ) {
-        items(
-            count = items.itemCount,
-            key = items.itemKey(itemKey),
-            contentType = items.itemContentType(itemContentType),
-        ) { position ->
-            items[position]?.let { item ->
-                itemSlots[item::class]?.let { slot -> slot(position, item) }
-            }
         }
     }
 }
